+++ pythonk/parser/internal/err_handling.py
@@ -1,4 +1,4 @@
-from pythonk.parser.internal.standard_lib import StandardGrammar
+from pythonk.parser.internal.standard_grammar import StandardGrammar
 
 
 class ErrorHandling(StandardGrammar):
